Update new order and adjust position tips
