+++ components/Portfolio/Trading/Position/index.tsx
@@ -112,7 +112,7 @@ const Position: React.FC<{
                         const name = tracer.marketId.split('/')[0];
                         const status = _status[i];
                         const { quote, base, totalMargin } = tracer.balances;
-                        const realisedPNL = 0; // TODO calculte realisedPNL
+                        const realisedPNL = 0; // TODO calculate realisedPNL
                         const unrealisedPNL = calcUnrealised(
                             base,
                             tracer.oraclePrice,
